feat!: migrate to 0-indexed grid indices and quantics digits
Breaking change (0.1.0 -> 0.2.0): the crate now uses 0-indexed grid
indices and quantics digits throughout, matching the rest of the
tensor4all-rs workspace and removing the only 1-indexed surface in it
(tensor4all/tensor4all-rs#584).

- Grid indices: 0..N (first grid point [0, ...]) instead of 1..=N.
- Quantics digits: 0/1 bit values (0..base-1 for non-binary bases)
  instead of 1/2.
- Index-table bit numbers (QuanticsIndex): 0-indexed (0 = MSB).
- Grid-index and quantics parameters/results are typed usize; original
  coordinates (origin/step/origcoord) stay i64/f64.
- max_grididx() now returns the maximum 0-based index (base^R - 1).
- Conversion formulas drop the +/-1: origcoord = origin + grididx*step.

Porting note for QuanticsGrids.jl scripts: subtract 1 at the call
boundary. Tests, doctests, doc examples, the mdBook guide, and the
benchmark all assert the 0-indexed semantics.

Closes tensor4all/quanticsgrids-rs#2.

+++ src/error.rs
@@ -82,25 +82,25 @@ pub enum QuanticsGridError {
     },
 
     /// Quantics value out of range
-    #[error("Quantics value {value} for site {site} out of range [1, {max}]")]
+    #[error("Quantics value {value} for site {site} out of range [0, {max}]")]
     QuanticsOutOfRange {
         /// Site index
         site: usize,
         /// The out-of-range value
-        value: i64,
+        value: usize,
         /// Maximum allowed value
-        max: i64,
+        max: usize,
     },
 
     /// Grid index out of bounds
-    #[error("Grid index {value} for dimension {dim} out of bounds [1, {max}]")]
+    #[error("Grid index {value} for dimension {dim} out of bounds [0, {max}]")]
     GridIndexOutOfBounds {
         /// Dimension index
         dim: usize,
         /// The out-of-bounds value
-        value: i64,
+        value: usize,
         /// Maximum allowed value
-        max: i64,
+        max: usize,
     },
 
     /// Coordinate out of bounds
